more things added back in
