+++ .gitignore
@@ -6,12 +6,8 @@ dist/
 lib/
 
 # Generated files
-/out/
 /doc/
 /gen/
-/gen3/
-/gen4/
-/tool/playground/
 tmp/
 
 # Configurable build files
