--- compiler/Test/edu/clemson/resolve/TestTopologicalSort.java
@@ -1,10 +1,0 @@
-package edu.clemson.resolve;
-
-import org.junit.Test;
-
-public class TestTopologicalSort {
-
-    @Test public void testMediumGraph() {
-
-    }
-}
