--- compiler/src/edu/clemson/resolve/semantics/TypedSymbol.java
@@ -1,6 +1,0 @@
-package edu.clemson.resolve.semantics;
-
-public interface TypedSymbol {
-	Type getType();
-	void setType(Type type);
-}
