+++ src/org/antlr/symbols/TypedSymbol.java
@@ -1,0 +1,6 @@
+package org.antlr.symbols;
+
+public interface TypedSymbol {
+	Type getType();
+	void setType(Type type);
+}
